// fe/src/app/models/player.ts
export interface Player {
    id: number
    resourceId: any
    resourceBaseId: any
    futBinId: any
    futWizId: any
    firstName: string
    lastName: string
    name: string
    commonName: string
    height: number
    weight: number
    gender: string
    birthDate: string
    age: number
    league: number
    nation: number
    club: number
    rarity: number
    playStyles: any
    playStylesPlus: any
    position: string
    positionAlternatives: any
    skillMoves: number
    weakFoot: number
    foot: string
    attackWorkRate: string
    defenseWorkRate: string
    totalStats: number
    totalStatsInGame: number
    color: string
    rating: number
    ratingAverage: number
    pace: number
    shooting: number
    passing: number
    dribbling: number
    defending: number
    physicality: number
    paceAttributes: any
    shootingAttributes: any
    passingAttributes: any
    dribblingAttributes: any
    defendingAttributes: any
    physicalityAttributes: any
    goalkeeperAttributes: any
    playerImage: any
    nationImage: any
    clubImage: any
  }